+++ .github/workflows/CI-python.yml
@@ -64,7 +64,7 @@ jobs:
     - name: Test with pytest
       shell: bash -l {0}
       run: |
-        pytest ./tests -s -v --durations=10 --cov='ml_wrappers' --cov-report=xml --cov-report=html
+        pytest ./tests/main -s -v --durations=10 --cov='ml_wrappers' --cov-report=xml --cov-report=html
     - name: Upload code coverage results
       uses: actions/upload-artifact@v3
       with:
