+++ tests/automl/test_automl_image_model_wrapper.py
@@ -10,29 +10,18 @@ import os
 import sys
 import tempfile
 
-import numpy as np
-import pandas as pd
+import azureml.automl.core.shared.constants as shared_constants
+import mlflow
 import pytest
 import torch
-
-try:
-    import azureml.automl.core.shared.constants as shared_constants
-    import mlflow
-    from azureml.automl.dnn.vision.classification.common.constants import \
-        ModelNames
-    from azureml.automl.dnn.vision.classification.models import ModelFactory
-    from azureml.automl.dnn.vision.common.mlflow.mlflow_model_wrapper import \
-        MLFlowImagesModelWrapper
-    from azureml.automl.dnn.vision.common.model_export_utils import (
-        _get_mlflow_signature, _get_scoring_method)
-except ImportError:
-    pass
-from common_vision_utils import (IMAGE, create_image_classification_pipeline,
-                                 create_pytorch_image_model,
-                                 load_base64_images, load_fridge_dataset,
-                                 load_imagenet_dataset, load_images,
-                                 preprocess_imagenet_dataset,
-                                 retrieve_or_train_fridge_model)
+from azureml.automl.dnn.vision.classification.common.constants import \
+    ModelNames
+from azureml.automl.dnn.vision.classification.models import ModelFactory
+from azureml.automl.dnn.vision.common.mlflow.mlflow_model_wrapper import \
+    MLFlowImagesModelWrapper
+from azureml.automl.dnn.vision.common.model_export_utils import (
+    _get_mlflow_signature, _get_scoring_method)
+from common_vision_utils import load_base64_images, load_fridge_dataset
 from ml_wrappers import wrap_model
 from ml_wrappers.common.constants import ModelTask
 from wrapper_validator import validate_wrapped_classification_model
@@ -40,35 +29,11 @@ from wrapper_validator import validate_wrapped_classification_model
 
 @pytest.mark.usefixtures('_clean_dir')
 class TestImageModelWrapper(object):
-    def test_wrap_resnet_classification_model(self):
-        data = load_imagenet_dataset()
-        pred = create_image_classification_pipeline()
-        wrapped_model = wrap_model(pred, data, ModelTask.IMAGE_CLASSIFICATION)
-        validate_wrapped_classification_model(wrapped_model, data)
-
-    # Skip for older versions of python due to many breaking changes in fastai
-    @pytest.mark.skipif(sys.version_info.minor <= 6,
-                        reason='Fastai not supported for older versions')
-    def test_wrap_fastai_image_classification_model(self):
-        data = load_fridge_dataset()
-        try:
-            model = retrieve_or_train_fridge_model(data)
-        except Exception as e:
-            print("Failed to retrieve or load Fastai model, force training")
-            print("Inner exception message on retrieving model: {}".format(e))
-            model = retrieve_or_train_fridge_model(data, force_train=True)
-        # load the paths as numpy arrays
-        data = load_images(data)
-        wrapped_model = wrap_model(model, data, ModelTask.IMAGE_CLASSIFICATION)
-        validate_wrapped_classification_model(wrapped_model, data)
-
     # Skip for older versions of python as azureml-automl-dnn-vision works with ">=3.7,<3.8"
     @pytest.mark.skipif(sys.version_info < (3, 7),
                         reason='azureml-automl-dnn-vision not supported for older versions')
     @pytest.mark.skipif(sys.version_info >= (3, 8),
                         reason='azureml-automl-dnn-vision not supported for newer versions')
-    @pytest.mark.skipif('azureml-automl-dnn-vision' not in sys.modules,
-                        reason="requires the azureml-automl-dnn-vision library")
     def test_wrap_automl_image_classification_model(self):
         data = load_fridge_dataset()
         model_name = ModelNames.SERESNEXT
@@ -135,32 +100,3 @@ class TestImageModelWrapper(object):
             wrapped_model = wrap_model(
                 mlflow_model, data, ModelTask.IMAGE_CLASSIFICATION)
             validate_wrapped_classification_model(wrapped_model, data)
-
-    # Skip for older versions of pytorch due to missing classes
-    @pytest.mark.skipif(sys.version_info.minor <= 6,
-                        reason='Older versions of pytorch not supported')
-    def test_pytorch_image_classification_model(self):
-        data = load_imagenet_dataset()[:3]
-        data = preprocess_imagenet_dataset(data)
-        model = create_pytorch_image_model()
-        wrapped_model = wrap_model(model, data, ModelTask.IMAGE_CLASSIFICATION)
-        validate_wrapped_classification_model(wrapped_model, data)
-
-    # Skip for older versions of pytorch due to missing classes
-    @pytest.mark.skipif(sys.version_info.minor <= 6,
-                        reason='Older versions of pytorch not supported')
-    def test_pytorch_image_classification_model_pandas(self):
-        data = load_imagenet_dataset()[:3]
-        data = preprocess_imagenet_dataset(data)
-        # convert to pandas dataframe of images
-        data = np.array(data)
-        # change back to channel at last dimension for
-        # standard image representation
-        data = np.moveaxis(data, 1, -1)
-        imgs = np.empty((len(data)), dtype=object)
-        for i, row in enumerate(data):
-            imgs[i] = row
-        data = pd.DataFrame(imgs, columns=[IMAGE])
-        model = create_pytorch_image_model()
-        wrapped_model = wrap_model(model, data, ModelTask.IMAGE_CLASSIFICATION)
-        validate_wrapped_classification_model(wrapped_model, data)
