+++ tests/main/test_text_model_wrapper.py
